server/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.7
gunicorn>=20.1.0
pytz>=2023.3
django-allauth>=0.54.0
django-cors-headers>=3.13.0
bcrypt>=4.0.1
finnhub-python>=2.4.0
pytest>=8.3.2
pytest-django>=4.5.2
factory_boy>=3.3.0
Faker>=19.10.0
flake8>=6.1.0
black>=23.9.1
flake8-django>=1.4
mysqlclient>=2.1
djangorestframework>=3.14.0
pytest-mock>=3.10.0
djangorestframework-simplejwt
